refactor(mb-api): fold tauxProgression module into valeurAvancement
Suite review PR #2210 : tauxProgression est un sous-domaine de valeurAvancement
(série + objectif + resolver pur), pas un module à part. Fusion du resolver,
de la query et de la route dans le module existant ; suppression du path
tsconfig dédié.
