+++ apps/mb-api/src/valeurAvancement/queries/listTauxProgressionForIndicateur.ts
@@ -21,7 +21,7 @@ import {
   type ObjectifBrut,
   resolveTauxProgression,
   type ValeurBrute,
-} from '@/tauxProgression/resolveTauxProgression'
+} from '@/valeurAvancement/resolveTauxProgression'
 import { loadResolveSerieContext } from '@/valeurAvancement/queries/loadResolveSerieContext'
 import {
   type IndividuRef,
