+++ apps/mb-api/src/valeurAvancement/resolveTauxProgression.ts
